Count path multiplicity using a long

+++ checkouts/java/src/main/java/nl/mauritssilvis/darts/checkouts/java/paths/GroupedPath.java
@@ -49,7 +49,7 @@ public class GroupedPath implements Path {
     }
 
     @Override
-    public int getMultiplicity() {
+    public long getMultiplicity() {
         return -1;
     }
 }
