+++ checkouts/java/src/main/java/nl/mauritssilvis/darts/checkouts/java/paths/Path.java
@@ -60,5 +60,5 @@ public interface Path {
      *
      * @return the multiplicity of this path
      */
-    int getMultiplicity();
+    long getMultiplicity();
 }
